requirements.txt
faiss-cpu>=1.13.1
sentence-transformers==3.2.1
pypdf==5.1.0
python-docx==1.1.2
pyyaml==6.0.2
llama-cpp-python==0.3.2
colorama==0.4.6
gradio>=4.0.0
